// packages/sdk/src/gen/outbox/v1/template_pb.d.ts
export declare type CreateTemplateRequest = Message<"outbox.v1.CreateTemplateRequest"> & {
  /**
   * @generated from field: string parent = 1;
   */
  parent: string;

  /**
   * @generated from field: outbox.v1.Template template = 2;
   */
  template?: Template;
};

/**
 * Describes the message outbox.v1.CreateTemplateRequest.
 * Use `create(CreateTemplateRequestSchema)` to create a new message.
 */
export declare const CreateTemplateRequestSchema: GenMessage<CreateTemplateRequest>;

/**
 * @generated from message outbox.v1.CreateTemplateResponse
 */
export declare type CreateTemplateResponse = Message<"outbox.v1.CreateTemplateResponse"> & {
  /**
   * @generated from field: outbox.v1.Template template = 1;
   */
  template?: Template;
};

/**
 * Describes the message outbox.v1.CreateTemplateResponse.
 * Use `create(CreateTemplateResponseSchema)` to create a new message.
 */
export declare const CreateTemplateResponseSchema: GenMessage<CreateTemplateResponse>;

/**
 * @generated from message outbox.v1.GetTemplateRequest
 */
export declare type GetTemplateRequest = Message<"outbox.v1.GetTemplateRequest"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;
};

/**
 * Describes the message outbox.v1.GetTemplateRequest.
 * Use `create(GetTemplateRequestSchema)` to create a new message.
 */
export declare const GetTemplateRequestSchema: GenMessage<GetTemplateRequest>;

/**
 * @generated from message outbox.v1.GetTemplateResponse
 */
export declare type GetTemplateResponse = Message<"outbox.v1.GetTemplateResponse"> & {
  /**
   * @generated from field: outbox.v1.Template template = 1;
   */
  template?: Template;
};

/**
 * Describes the message outbox.v1.GetTemplateResponse.
 * Use `create(GetTemplateResponseSchema)` to create a new message.
 */
export declare const GetTemplateResponseSchema: GenMessage<GetTemplateResponse>;

/**
 * @generated from message outbox.v1.ListTemplatesRequest
 */
export declare type ListTemplatesRequest = Message<"outbox.v1.ListTemplatesRequest"> & {
  /**
   * @generated from field: string parent = 1;
   */
  parent: string;

  /**
   * @generated from field: int32 page_size = 2;
   */
  pageSize: number;

  /**
   * @generated from field: string page_token = 3;
   */
  pageToken: string;
};

/**
 * Describes the message outbox.v1.ListTemplatesRequest.
 * Use `create(ListTemplatesRequestSchema)` to create a new message.
 */
export declare const ListTemplatesRequestSchema: GenMessage<ListTemplatesRequest>;

/**
 * @generated from message outbox.v1.ListTemplatesResponse
 */
export declare type ListTemplatesResponse = Message<"outbox.v1.ListTemplatesResponse"> & {
  /**
   * @generated from field: repeated outbox.v1.Template templates = 1;
   */
  templates: Template[];

  /**
   * @generated from field: string next_page_token = 2;
   */
  nextPageToken: string;

  /**
   * @generated from field: int32 total_size = 3;
   */
  totalSize: number;
};

/**
 * Describes the message outbox.v1.ListTemplatesResponse.
 * Use `create(ListTemplatesResponseSchema)` to create a new message.
 */
export declare const ListTemplatesResponseSchema: GenMessage<ListTemplatesResponse>;

/**
 * @generated from message outbox.v1.DeleteTemplateRequest
 */
export declare type DeleteTemplateRequest = Message<"outbox.v1.DeleteTemplateRequest"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;
};

/**
 * Describes the message outbox.v1.DeleteTemplateRequest.
 * Use `create(DeleteTemplateRequestSchema)` to create a new message.
 */
export declare const DeleteTemplateRequestSchema: GenMessage<DeleteTemplateRequest>;

/**
 * @generated from message outbox.v1.DeleteTemplateResponse
 */
export declare type DeleteTemplateResponse = Message<"outbox.v1.DeleteTemplateResponse"> & {
};

/**
 * Describes the message outbox.v1.DeleteTemplateResponse.
 * Use `create(DeleteTemplateResponseSchema)` to create a new message.
 */
export declare const DeleteTemplateResponseSchema: GenMessage<DeleteTemplateResponse>;
// ...
export declare const TemplateService: GenService<{
  /**
   * @generated from rpc outbox.v1.TemplateService.CreateTemplate
   */
  createTemplate: {
    methodKind: "unary";
    input: typeof CreateTemplateRequestSchema;
    output: typeof CreateTemplateResponseSchema;
  },
  /**
   * @generated from rpc outbox.v1.TemplateService.GetTemplate
   */
  getTemplate: {
    methodKind: "unary";
    input: typeof GetTemplateRequestSchema;
    output: typeof GetTemplateResponseSchema;
  },
  /**
   * @generated from rpc outbox.v1.TemplateService.ListTemplates
   */
  listTemplates: {
    methodKind: "unary";
    input: typeof ListTemplatesRequestSchema;
    output: typeof ListTemplatesResponseSchema;
  },
  /**
   * @generated from rpc outbox.v1.TemplateService.DeleteTemplate
   */
  deleteTemplate: {
    methodKind: "unary";
    input: typeof DeleteTemplateRequestSchema;
    output: typeof DeleteTemplateResponseSchema;
  },
}>;